uadate file

+++ server/routes/post.route.js
@@ -49,14 +49,16 @@ postrouter.put("/:id", async (req, res) => {
 //delete a post
 
 postrouter.delete("/:id", async (req, res) => {
+
+  // console.log(req.body.userId);
   try {
     const post = await Post.findById(req.params.id);
-    // if (post.userId === req.body.userId) {
+    if (post.userId === req.body.userId) {
       await post.deleteOne();
       res.status(200).json("the post has been deleted");
-    // } else {
-    //   res.status(403).json("you can delete only your post");
-    // }
+    } else {
+      res.status(403).json("you can delete only your post");
+    }
   } catch (err) {
     res.status(500).json(err);
   }
